data/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-cloud-bigquery==3.27.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.22.0
google-cloud-compute==1.23.0
pandas==2.2.3
pyarrow==22.0.0
db-dtypes==1.4.4
pandas-gbq==0.25.0
colorlog==6.9.0
yfinance==0.2.59
SQLAlchemy==2.0.38
cloud-sql-python-connector==1.16.0
pg8000==1.31.2
google-cloud-storage==3.0.0
dbt-core==1.10.8
dbt-bigquery==1.10.2
apache-airflow-providers-google==19.2.0
sqlfluff==3.5.0
dbt-colibri==0.2.7
playwright==1.49.0
lxml==6.0.2
